Extract the main display out of the details display so it can be toggled with more data. Also add buttons to toggle and to close the display.
