+++ src/components/DetailsMain/DetailsMain.scss
@@ -1,0 +1,67 @@
+@import "../../sass/colors";
+@import "../../sass/dimensions";
+@import "../../sass/fonts";
+
+.launch-details {
+  position: fixed;
+  width: 80%;
+  max-width: 500px;
+  height: auto;
+  background: linear-gradient(
+    $color-details-main-light,
+    $color-details-main-dark
+  );
+  padding: 20px;
+  border-radius: 15px;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-between;
+
+  p {
+    margin: 0;
+  }
+}
+
+.success {
+  color: $color-success;
+}
+
+.fail {
+  color: $color-fail;
+}
+
+.button-display {
+  display: flex;
+  justify-content: center;
+  column-gap: $button-gap;
+  width: 100%;
+  margin-top: 20px;
+}
+
+.btn-set {
+  font-size: $font-small;
+  font-weight: 700;
+  outline: none;
+  border: none;
+  cursor: pointer;
+
+  &:hover {
+    opacity: 0.7;
+  }
+}
+
+#more-button {
+  padding: $padding-details-text-small $padding-details-text-medium;
+  border-radius: $radius-details-section;
+  background-color: $color-success;
+  color: $color-font-dark;
+}
+
+#close-display {
+  background-color: transparent;
+  color: $color-font-light;
+}
