+++ src/sass/_dimensions.scss
@@ -2,3 +2,5 @@ $padding-details-text-small: 5px;
 $padding-details-text-medium: 10px;
 
 $radius-details-section: 5px;
+
+$button-gap: 25px;
